requirements.txt
opencv_python==4.2.0.34
# joblib==0.16.0
Keras==2.3.1
ipinfo==3.0.0
h5py==2.10.0
PassportEye==1.4.1
pandas==1.0.3
matplotlib==3.1.1
requests==2.23.0
numpy==1.18.1
tensorflow==2.2.0
deepface==0.0.22
Flask==1.1.2
pytesseract==0.3.4
Pillow==7.1.2
progressbar33==2.4
scikit_learn==0.23.1
cmake==3.17.3
dlib==19.19.0
ply==3.11
jsonpickle==1.4.1
Flask-Cors==3.0.8
python-dotenv==0.13.0
